+++ src/main/java/entities/Expense.java
@@ -1,4 +1,4 @@
-package main.entities;
+package entities;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
