+++ src/main/java/entities/ExpenseFactory.java
@@ -1,4 +1,4 @@
-package main.entities;
+package entities;
 
 import java.util.ArrayList;
