+++ src/main/java/entities/Group.java
@@ -1,4 +1,4 @@
-package main.entities;
+package entities;
 
 import java.util.ArrayList;
 import java.util.List;
