+++ src/main/java/usecase/SettleUp/SettleUpOutputBoundary.java
@@ -1,4 +1,4 @@
-package main.usecase;
+package usecase.SettleUp;
 
 public interface SettleUpOutputBoundary {
     void present(SettleUpOutputData outputData);
