+++ src/main/java/view/MainFrame.java
@@ -1,4 +1,4 @@
-package main.view;
+package view;
 
 import javax.swing.*;
